requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
azure-cognitiveservices-search-websearch==2.0.1
azure-common==1.1.28
azure-core==1.30.2
azure-mgmt-core==1.4.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
boto3==1.34.119
botocore==1.34.162
bs4==0.0.2
build==1.2.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
cymem==2.0.8
dataclasses-json==0.6.7
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fastapi==0.114.0
filelock==3.16.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
greenlet==3.0.3
groq==0.9.0
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.4
isodate==0.6.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kubernetes==30.1.0
langchain==0.2.7
langchain-chroma==0.1.1
langchain-community==0.2.7
langchain-core==0.2.38
langchain-google-genai==1.0.10
langchain-groq==0.1.9
langchain-openai==0.1.10
langchain-pinecone==0.1.1
langchain-text-splitters==0.2.4
langcodes==3.4.0
langsmith==0.1.116
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msrest==0.7.1
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
narwhals==1.6.3
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.44.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pillow==10.4.0
pinecone-client==3.2.2
posthog==3.6.3
preshed==3.0.9
proto-plus==1.24.0
protobuf==4.25.4
psycopg==3.1.18
psycopg-binary==3.1.19
psycopg-pool==3.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.9.0
pydantic_core==2.23.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.13.1
sentence-transformers==2.7.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.34
srsly==2.4.8
starlette==0.38.5
streamlit==1.38.0
sympy==1.13.2
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.2
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.30.6
wasabi==1.1.3
watchdog==4.0.2
watchfiles==0.24.0
weasel==0.3.4
websocket-client==1.8.0
websockets==13.0.1
wrapt==1.16.0
yarl==1.11.0
youtube-transcript-api==0.6.2
zipp==3.20.1
